pyproject.toml
[tool.poetry]
name = "decouple-extended"
version = "0.1.5"
description = " Simple extensions to python-decouple"
authors = ["Roberto Alamos <ralamosm@gmail.com>"]
readme = "README.md"
packages = [{include = "decouple_extended"}]

[tool.poetry.dependencies]
python = "^3.8"
python-decouple = "^3.8"
boto3 = "^1.28.33"
pydantic = "^1.10.11"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pre-commit = "^3.3.3"
moto = "^4.2.0"

[tool.poetry.group.test]
optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
